test(fetch): split test_guarded_fetcher by subject
The file mixed three subjects behind one transport. Split by what each test
actually exercises, and deleted the transport's own matrix along with the
class it drove.

Deleted (drove ``GuardedFetcher`` itself): happy-path/IP-pinning, lowest-
address pinning, scheme and URL-shape refusals through the fetcher, the http
allowlist through the fetcher, redirect re-validation and the hop cap,
streamed byte-cap and per-category cap, digest mismatch and malformed digest,
injector/policy, Host header, "the resolver runs again on every fetch",
connection errors, non-2xx status.

Moved to ``fetch/test_endpoint_guard.py``, rewritten to call
``endpoint_refusal`` directly with an injected resolver: non-global addresses
refused, one private among public answers refuses the host, the allowlist
exempts address validation (and does not cover other hosts), and the
unresolvable-host case — which at write time is logged and stored rather than
refused, so the test now pins that divergence and the fact that the raising
resolver does not escape. Added the literal-address case, since a numeric
host needs no DNS and must still be refused.

Moved to ``fetch/test_limits.py`` (new): the four allowlist parsing tests and
the category/cap-table binding test.

``test_manifest_fetch_module_wiring.py``'s allowlist-wiring test is untouched
— it is the test that proves the config key is still alive — as are the
golden config JSONs.

Also drops the last docstring reference to the deleted ``AuthorizationPolicy``
Protocol from ``PrefixAuthorizationPolicy``, which is alive and stays.
